Move app-level enhancement into the theme
