+++ .vuepress/theme/enhanceApp.js
@@ -1,11 +1,10 @@
-
-import AuthorList from './theme/AuthorList.vue'
+import AuthorList from '@theme/components/AuthorList.vue'
 
 export default ({
   Vue, // the version of Vue being used in the VuePress app
   options, // the options for the root Vue instance
   router, // the router instance for the app
-  siteData // site metadata
+  siteData, // site metadata
 }) => {
-  Vue.component('author-list', AuthorList)
+  Vue.component('AuthorList', AuthorList)
 }
